pyproject.toml
[project]
name = "alma-n8n-project-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.31.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.1",
]
